refactor: correct code for clean
